rtvi: separate specific google RTVI into a GoogleRTVIObserver
